Fix issue where new service account ENV var is forced in every process
ManageIQ::Providers::Workflows::Engine.floe_docker_runner is called in a
config/initializer, which means that it is being executed in every Rails
environment on boot. This includes things like `bin/rake evm:deployment_status`,
which happens very early on in the orchestrator boot process.

When the orchestrator is on Kubernetes, it starts creating a
Floe::Kubernetes::Runner, but then requires the new
AUTOMATION_JOB_SERVICE_ACCOUNT, which doesn't exist as it is only meant
to exist in the new AutomationWorker. Because of this the orchestrator
fails to boot.

This commit allows that value to be nil, so it can move on. However, we
should find a way to avoid creating the runner on every boot, which is a
better solution.