Fix Logger error when called from service
When called from a MiqRequest object we use a
`Vmdb::Loggers::RequestLogger` so that floe logs can be seen in the UI
easily.  If `Floe.logger` is already a `BroadcastLogger` then
`Floe.logger.wrap` isn't calling `ManageIQ::Loggers::Base#wrap` it is
actually calling `ActiveSupport::BroadcastLogger#method_missing` which
does `loggers.map { |logger| logger.send(name, ...) }` and results in an
array being returned.

This array is then used as the `workflow.context.logger` so when we try
to log to it we get `undefined method info for Array`